Remove _sponsors folder and sponsors.md
This section has to be re-worked. I think we should table it for now so
that the new graphics can be published. When its ready to go back in,
just create the same folder _sponsors and the same sponsors.md file
(file has no content). The supporting css and twig files are still in
the repo.